Move test utility method up in class inheritance
